Restore WebSocket endpoint
